+++ packages/eslint-plugin-isaacscript/src/configs/recommended.ts
@@ -40,6 +40,7 @@ export const recommended = {
     "isaacscript/require-variadic-function-argument": "error",
     "isaacscript/strict-array-methods": "error",
     "isaacscript/strict-enums": "error",
+    "isaacscript/strict-undefined-functions": "error",
     "isaacscript/strict-void-functions": "error",
   },
 };